+++ src/briefy/leica/models/job/__init__.py
@@ -1,0 +1,1 @@
+"""Assignment, Order, LeadOrder, OrderLocation and Pool models."""