+++ src/briefy/leica/subscribers/professional.py
@@ -21,3 +21,20 @@ def professional_created_handler(event):
 
     # submit for QA validation after creation
     obj.workflow.submit()
+
+
+def professional_ownership(event):
+    """Make sure Professional user is owner of it is own profile."""
+    professional = event.obj
+    professional_id = professional.id
+    owner = professional.owner
+    if str(professional_id) != str(owner):
+        professional.owner = professional_id
+
+
+def transition_handler(event):
+    """Handle Professional transition events."""
+    event_name = event.event_name
+    if not event_name.startswith('professional.workflow'):
+        return
+    professional_ownership(event)