+++ tests/views/test_orders_view.py
@@ -65,3 +65,16 @@ class TestOrderView(BaseVersionedTestView):
         assert error['name'] == 'asset_types'
         assert error['location'] == 'body'
         assert 'Invalid number of type of assets' in error['description']
+
+    def test_unsuccessful_creation(self, obj_payload, app):
+        """Test unsuccessful creation of a new model."""
+        payload = obj_payload
+        payload['id'] = 'e93b5902-c15e-4b47-8e01-a93df6ea7211'
+        # Use a project that does not allow creation of new orders
+        payload['project_id'] = '4a068a1b-3646-4acf-937d-15563853e388'
+        request = app.post_json(self.base_path, payload, headers=self.headers, status=403)
+
+        assert request.status_code == 403
+        result = request.json
+        assert result['status'] == 'error'
+        assert result['message'] == 'Unauthorized'